Added more todos and some minor changes.
