+++ main/src/ca/uwaterloo/flix/language/phase/jvm/VariableInfo.scala
@@ -5,6 +5,7 @@ package ca.uwaterloo.flix.language.phase.jvm
   */
 sealed trait VariableInfo
 
+// TODO: Magnus: Why is this not used?
 object VariableInfo {
 
   /**
